+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Fix the incompatiblity with 3.13 because of the `datetime.replace()` change. ([#333](https://github.com/python-poetry/tomlkit/issues/333))
 - Revert the change of parsing out-of-order tables. ([#347](https://github.com/python-poetry/tomlkit/issues/347))
+- Keep the nested out-of-order table. ([#361](https://github.com/python-poetry/tomlkit/issues/361))
 
 
 ## [0.12.5] - 2024-05-08